Fix error converting color into string

+++ gramps/gui/undohistory.py
@@ -254,9 +254,9 @@ def gdk_color_to_str(color):
     """
     Convert a Gdk.Color into a #rrggbb string.
     """
-    color_str = "#%02x%02x%02x" % (color.red * 255,
-                                    color.green * 255,
-                                    color.blue * 255)
+    color_str = "#%02x%02x%02x" % (int(color.red * 255),
+                                   int(color.green * 255),
+                                   int(color.blue * 255))
     return color_str
 
 def get_colors(obj, state):
